+++ sharding/nodesCoordinator/sovereignIndexHashedNodesCoordinator_test.go
@@ -122,61 +122,3 @@ func TestSovereignIndexHashedNodesCoordinator_GetConsensusValidatorsPublicKeys(t
 		require.Nil(t, err)
 	})
 }
-
-func TestSovereignIndexHashedNodesCoordinator_ComputeNumberOfShards(t *testing.T) {
-	t.Parallel()
-
-	t.Run("should work only with 1", func(t *testing.T) {
-		t.Parallel()
-
-		arguments := createArguments()
-		sihnc, _ := NewSovereignIndexHashedNodesCoordinator(arguments)
-
-		eligibleMap := make(map[uint32][]Validator)
-		eligibleMap[0] = []Validator{&validator{}}
-
-		nodesConfig := &epochNodesConfig{
-			eligibleMap: eligibleMap,
-		}
-		nbShards, err := sihnc.ComputeNumberOfShards(nodesConfig)
-
-		require.Nil(t, err)
-		require.Equal(t, uint32(1), nbShards)
-	})
-
-	t.Run("should not work with 0", func(t *testing.T) {
-		t.Parallel()
-
-		arguments := createArguments()
-		ihnc, _ := NewSovereignIndexHashedNodesCoordinator(arguments)
-
-		eligibleMap := make(map[uint32][]Validator)
-
-		nodesConfig := &epochNodesConfig{
-			eligibleMap: eligibleMap,
-		}
-		nbShards, err := ihnc.ComputeNumberOfShards(nodesConfig)
-
-		require.NotNil(t, err)
-		require.Equal(t, uint32(0), nbShards)
-	})
-
-	t.Run("should not work with 2", func(t *testing.T) {
-		t.Parallel()
-
-		arguments := createArguments()
-		ihnc, _ := NewSovereignIndexHashedNodesCoordinator(arguments)
-
-		eligibleMap := make(map[uint32][]Validator)
-		eligibleMap[0] = []Validator{&validator{}}
-		eligibleMap[1] = []Validator{&validator{}}
-
-		nodesConfig := &epochNodesConfig{
-			eligibleMap: eligibleMap,
-		}
-		nbShards, err := ihnc.ComputeNumberOfShards(nodesConfig)
-
-		require.NotNil(t, err)
-		require.Equal(t, uint32(0), nbShards)
-	})
-}
